+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Auto/PixelDropoffRed.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.Auto;
 
 import android.util.Size;
